changed use of va_list ap to va_list *ap

+++ accufield_functions.c
@@ -6,14 +6,14 @@
 /*   By: anleclab <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/29 17:41:20 by anleclab          #+#    #+#             */
-/*   Updated: 2018/12/03 16:01:58 by anleclab         ###   ########.fr       */
+/*   Updated: 2018/12/03 16:30:58 by anleclab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 #include "libft/libft.h"
 
-char	*accuracy(char *str, int len, char c)
+char	*accuracy(char *str, size_t len, char c)
 {
 	char	*ret;
 	int		i;
@@ -60,7 +60,7 @@ char	*accuracy(char *str, int len, char c)
 	return (ret);
 }
 
-char	*field_width(char *str, int len)
+char	*field_width(char *str, size_t len)
 {
 	char	*ret;
 	int		i;
